Restore explicit exit code handling
This commit restores the error handling behavior for the
ListenAndServe code block to allow return of an explicit
error code to the OS (i.e., the process manager)

This behavior was restored by using an initial defer of a
pointer to int with a default value that indicates success.
This can be overridden as needed (and currently is)
to indicate application failure without blocking other
deferred functions from running.

refs GH-179