Add support for limiting payloads to specific IPs
If provided, check remote IP Address against the list
of user-provided IP Addresses. If a match is not found,
reject the connection with forbidden status. If a match
is found, accept the connection.

If the sysadmin does not provide a list of trusted IP Addresses
via config file or CLI flag, accept payloads from any IP
Address. An attempt to balance this choice is provided by
setting the default value in the starter config file to
127.0.0.1, or "localhost".

This behavior is subject to change; the behavior may change
to require an explicit list of IP Addresses to receive
payloads.

Update docs, config template to reflect the changes.

refs GH-18